Use prescaler to set the frequency
Earlier, we were changing the ARR value according to required
frequency. Now we are using the prescaler value, it also says in the
documentation that it is buffered and can be used to change timer
frequency "on the fly".

Results are somehow worse? The tone, whatever part of it that play,
sounds somehow off, and actually, lesser of it plays, than before.